feat: adjusting repository page
